chore_: moving content of nwaku.go to nwaku_test_utils.go
